refactor - #111 user 패키지를 auth 패키지로 변경
