+++ src/main/java/sopt/org/hmh/domain/auth/controller/UserApi.java
@@ -1,4 +1,4 @@
-package sopt.org.hmh.domain.user.controller;
+package sopt.org.hmh.domain.auth.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -7,8 +7,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
-import sopt.org.hmh.domain.user.dto.request.SocialPlatformRequest;
-import sopt.org.hmh.domain.user.dto.request.SocialSignUpRequest;
+import sopt.org.hmh.domain.auth.dto.request.SocialPlatformRequest;
+import sopt.org.hmh.domain.auth.dto.request.SocialSignUpRequest;
 import sopt.org.hmh.global.auth.UserId;
 import sopt.org.hmh.global.auth.jwt.JwtConstants;
 import sopt.org.hmh.global.common.response.BaseResponse;
