+++ src/main/java/sopt/org/hmh/domain/auth/domain/OnboardingInfo.java
@@ -1,10 +1,9 @@
-package sopt.org.hmh.domain.user.domain;
+package sopt.org.hmh.domain.auth.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -12,20 +11,19 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @Entity
-@Table(name = "problem")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class OnboardingProblem {
+public class OnboardingInfo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long onboardingInfoId;
-    private String problem;
+    private String averageUseTime;
+    private Long userId;
 
     @Builder
-    public OnboardingProblem(Long onboardingInfoId, String problem) {
-        this.onboardingInfoId = onboardingInfoId;
-        this.problem = problem;
+    public OnboardingInfo(String averageUseTime, Long userId) {
+        this.averageUseTime = averageUseTime;
+        this.userId = userId;
     }
 }
