+++ src/main/java/sopt/org/hmh/domain/auth/dto/response/LoginResponse.java
@@ -1,7 +1,7 @@
-package sopt.org.hmh.domain.user.dto.response;
+package sopt.org.hmh.domain.auth.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import sopt.org.hmh.domain.user.domain.User;
+import sopt.org.hmh.domain.auth.domain.User;
 import sopt.org.hmh.global.auth.jwt.TokenResponse;
 
 public record LoginResponse(
