+++ src/main/java/sopt/org/hmh/domain/auth/repository/OnboardingInfoRepository.java
@@ -1,7 +1,7 @@
-package sopt.org.hmh.domain.user.repository;
+package sopt.org.hmh.domain.auth.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sopt.org.hmh.domain.user.domain.OnboardingInfo;
+import sopt.org.hmh.domain.auth.domain.OnboardingInfo;
 
 public interface OnboardingInfoRepository extends JpaRepository<OnboardingInfo, Long> {
 }
