+++ src/main/java/sopt/org/hmh/domain/auth/repository/ProblemRepository.java
@@ -1,7 +1,7 @@
-package sopt.org.hmh.domain.user.repository;
+package sopt.org.hmh.domain.auth.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sopt.org.hmh.domain.user.domain.OnboardingProblem;
+import sopt.org.hmh.domain.auth.domain.OnboardingProblem;
 
 public interface ProblemRepository extends JpaRepository<OnboardingProblem, Long> {
 }
