+++ src/main/java/sopt/org/hmh/domain/auth/service/UserService.java
@@ -1,4 +1,4 @@
-package sopt.org.hmh.domain.user.service;
+package sopt.org.hmh.domain.auth.service;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -7,20 +7,20 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import sopt.org.hmh.domain.challenge.service.ChallengeService;
-import sopt.org.hmh.domain.user.domain.OnboardingInfo;
-import sopt.org.hmh.domain.user.domain.OnboardingProblem;
-import sopt.org.hmh.domain.user.domain.User;
-import sopt.org.hmh.domain.user.domain.UserConstants;
-import sopt.org.hmh.domain.user.domain.exception.UserError;
-import sopt.org.hmh.domain.user.domain.exception.UserException;
-import sopt.org.hmh.domain.user.dto.request.SocialPlatformRequest;
-import sopt.org.hmh.domain.user.dto.request.SocialSignUpRequest;
-import sopt.org.hmh.domain.user.dto.response.LoginResponse;
-import sopt.org.hmh.domain.user.dto.response.ReissueResponse;
-import sopt.org.hmh.domain.user.dto.response.UserInfoResponse;
-import sopt.org.hmh.domain.user.repository.OnboardingInfoRepository;
-import sopt.org.hmh.domain.user.repository.ProblemRepository;
-import sopt.org.hmh.domain.user.repository.UserRepository;
+import sopt.org.hmh.domain.auth.domain.OnboardingInfo;
+import sopt.org.hmh.domain.auth.domain.OnboardingProblem;
+import sopt.org.hmh.domain.auth.domain.User;
+import sopt.org.hmh.domain.auth.domain.UserConstants;
+import sopt.org.hmh.domain.auth.domain.exception.UserError;
+import sopt.org.hmh.domain.auth.domain.exception.UserException;
+import sopt.org.hmh.domain.auth.dto.request.SocialPlatformRequest;
+import sopt.org.hmh.domain.auth.dto.request.SocialSignUpRequest;
+import sopt.org.hmh.domain.auth.dto.response.LoginResponse;
+import sopt.org.hmh.domain.auth.dto.response.ReissueResponse;
+import sopt.org.hmh.domain.auth.dto.response.UserInfoResponse;
+import sopt.org.hmh.domain.auth.repository.OnboardingInfoRepository;
+import sopt.org.hmh.domain.auth.repository.ProblemRepository;
+import sopt.org.hmh.domain.auth.repository.UserRepository;
 import sopt.org.hmh.global.auth.jwt.JwtProvider;
 import sopt.org.hmh.global.auth.jwt.JwtValidator;
 import sopt.org.hmh.global.auth.jwt.exception.JwtError;
